Fixed bundle install in appveyor
