+++ test/cookbooks/test/recipes/certificate.rb
@@ -107,21 +107,3 @@ windows_certificate 'C:/certs/test-cert.pfx' do
   pfx_password 'chef123'
   store_name 'CA'
 end
-
-# windows_certificate_binding 'ChefDummyCertForTest' do
-#   store_name 'CA'
-#   port 443
-# end
-
-# windows_certificate_binding '444-appid' do
-#   cert_name 'ChefDummyCertForTest'
-#   store_name 'CA'
-#   port 444
-#   app_id '{00000000-0000-0000-0000-000000000000}'
-# end
-
-# windows_certificate_binding '443-hostname' do
-#   cert_name 'ChefDummyCertForTest'
-#   store_name 'CA'
-#   address 'www.chef.io'
-# end
